Stage 1 of refactoring clean up
